backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
python-dotenv>=1.0.1
websockets>=12.0
openai>=1.35.0
pydantic>=2.7.4
